feat(structure): 修复一些Bug
