+++ internal/types/structure.go
@@ -31,7 +31,7 @@ type PutTeamNodeResp struct {
 
 // 新增团队，初始化团队架构（入参）
 type PostTeamReq struct {
-	Name string `json:"team_structures" binding:"required"`
+	Name string `json:"team_name" binding:"required"`
 }
 
 // 新增团队，初始化团队架构(出参）
